Module serato_tags::markers2

The Serato Markers2 tag stores various kinds of track "markers" like Cue Points, Saved Loops, Flips.

It also stores information about the tracks' color in the tracklist and if the track's beatgrid is locked.

Note that some of this information is also stored in Serato Markers_, and Serato will prefer that data over the information stored in Serato Markers2 if it is present.

The minimum length of this tag seems to be 470 bytes, and shorter contents are padded with null bytes.

Structs

BPMLockMarker
A BPMLOCK marker.

CensorFlipAction
A Censor action inside a FLIP marker.

CueMarker
A CUE marker.

FlipMarker
A FLIP marker.

JumpFlipAction
A Jump action inside a FLIP marker.

LoopMarker
A LOOP marker.

Markers2
Represents the Serato Markers2 tag.

Markers2Content
Represents the base64-encoded content of the Serato Markers2 tag.

TrackColorMarker
A COLOR marker.

UnknownFlipAction
An unknown FLIP action that we don't have a parser for.

UnknownMarker
An unknown marker that we don't have a parser for.

Enums

FlipAction
An action inside a FLIP marker.

Marker
A marker in the Serato Markers2 tag.

Functions

decode_base64_chunks

parse

parse_markers2_content

parse_utf8

take_base64_chunk

take_base64_chunks

take_bool

take_bpmlock_marker

take_color_marker

take_cue_marker

take_flip_marker

take_flip_marker_action

take_flip_marker_action_censor

take_flip_marker_action_jump

take_loop_marker

take_marker

take_marker_name

take_utf8