Module serato_tags::tag

Parsers for Serato's file tags

Re-exports

pub use container::TagContainer;

pub use container::TagType;

pub use self::analysis::Analysis;

pub use self::autotags::Autotags;

pub use self::beatgrid::Beatgrid;

pub use self::markers::Markers;

pub use self::markers2::Markers2;

pub use self::overview::Overview;

pub use self::relvolad::RelVolAd;

pub use self::vidassoc::VidAssoc;

Modules

analysis
The Serato Analysis tag stores the analysis version.

autotags
The Serato Autotags tag stores BPM and Gain values.

beatgrid
The Serato BeatGrid tag stores the beatgrid markers.

container
Provides the TagContainer struct, a simple interface to access data in Serato's tags.

format
Traits for the different tag types (ID3, FLAC, etc.)

generic
Generic structs

markers
The Serato Markers_ tag stores information about the first 5 Cues, 9 Loops and the track color.

markers2
The Serato Markers2 tag stores various kinds of track "markers" like Cue Points, Saved Loops, Flips.

overview
The Serato Overview tag stores the waveform overview data.

relvolad
The Serato RelVolAd tag stores the analysis version.

serato32
These functions convert between a custom 4-byte format (that we'll call serato32 for brevity) and 3-byte plaintext (both u32). Serato's custom format inserts a single null bit after every 7 payload bits, starting from the rightmost bit.

vidassoc
The Serato VidAssoc tag stores the analysis version.