Module serato_tags::tag::markers

The Serato Markers_ tag stores information about the first 5 Cues, 9 Loops and the track color.

This is redundant with some of the information from the Serato Markers2 tag. Serato will prefer information from Serato Markers_ if it's present.

Structs

Marker
Represents a single marker in the Serato Markers_ tag.

Markers
Represents the Serato Markers_ tag.

Enums

EntryType
The Type of a Marker.

Functions

take_bool
Returns a bool parsed from the next input byte.

take_position
Returns an Option<u32> which contains the position parsed from the next 5 input bytes.