Module serato_tags::util

Various helper utilities for simplify parsing.

Modules

serato32
These functions convert between a custom 4-byte format (that we'll call serato32 for brevity) and 3-byte plaintext (both u32). Serato's custom format inserts a single null bit after every 7 payload bits, starting from the rightmost bit.

Structs

Color
Represents a 3-Byte RGB color value.

Version
Represents 2-Byte version value.

Functions

take_color
Returns a Color struct parsed from the first 3 input bytes.

take_until_nullbyte
Returns the input slice until the first occurrence of a null byte.

take_version
Returns a Version struct parsed from the first 2 input bytes.